+++ packages/api-explorer/src/StandaloneApiExplorer.tsx
@@ -24,12 +24,12 @@
 
  */
 
-import React, { FC } from 'react'
-import { useRouteMatch } from 'react-router-dom'
+import React, { FC, useEffect, useState } from 'react'
 import {
   RunItProvider,
   defaultConfigurator,
   initRunItSdk,
+  loadSpecsFromVersions,
 } from '@looker/run-it'
 import { IAPIMethods } from '@looker/sdk-rtl'
 import { SpecList } from '@looker/sdk-codegen'
@@ -38,29 +38,41 @@ import { Provider } from 'react-redux'
 import ApiExplorer from './ApiExplorer'
 import { configureStore } from './state'
 import { StandaloneEnvAdaptor } from './utils'
+import { Loader } from './components'
 
-export interface StandloneApiExplorerProps {
-  specs: SpecList
+export interface StandaloneApiExplorerProps {
   headless?: boolean
+  versionsUrl: string
 }
 
 const standaloneEnvAdaptor = new StandaloneEnvAdaptor()
 const store = configureStore()
 
-export const StandaloneApiExplorer: FC<StandloneApiExplorerProps> = ({
-  specs,
+export const StandaloneApiExplorer: FC<StandaloneApiExplorerProps> = ({
   headless = false,
+  versionsUrl = '',
 }) => {
-  const match = useRouteMatch<{ specKey: string }>(`/:specKey`)
-  const specKey = match?.params.specKey || ''
-  // TODO we may not need this restriction any more?
-  // Check explicitly for specs 3.0 and 3.1 as run it is not supported.
-  // This is done as the return from OAUTH does not provide a spec key
-  // but an SDK is needed.
-  const chosenSdk: IAPIMethods | undefined =
-    specKey === '3.0' || specKey === '3.1'
-      ? undefined
-      : initRunItSdk(defaultConfigurator)
+  const [specs, setSpecs] = useState<SpecList | undefined>()
+  const [embedded, setEmbedded] = useState<boolean>(headless)
+  const [currentVersionsUrl, setCurrentVersionsUrl] =
+    useState<string>(versionsUrl)
+
+  useEffect(() => {
+    if (currentVersionsUrl) {
+      // Load specifications from the versions url
+      loadSpecsFromVersions(currentVersionsUrl).then((response) => {
+        setSpecs(response.specs)
+        if ('headless' in response) {
+          // headless will default to false if it's not explicitly set
+          setEmbedded(response.headless)
+        }
+      })
+    } else {
+      setSpecs(undefined)
+    }
+  }, [currentVersionsUrl])
+
+  const chosenSdk: IAPIMethods = initRunItSdk(defaultConfigurator)
 
   return (
     <Provider store={store}>
@@ -69,11 +81,18 @@ export const StandaloneApiExplorer: FC<StandloneApiExplorerProps> = ({
         configurator={defaultConfigurator}
         basePath="/api/4.0"
       >
-        <ApiExplorer
-          specs={specs}
-          envAdaptor={standaloneEnvAdaptor}
-          headless={headless}
-        />
+        <>
+          {specs ? (
+            <ApiExplorer
+              specs={specs}
+              envAdaptor={standaloneEnvAdaptor}
+              headless={embedded}
+              setVersionsUrl={setCurrentVersionsUrl}
+            />
+          ) : (
+            <Loader themeOverrides={standaloneEnvAdaptor.themeOverrides()} />
+          )}
+        </>
       </RunItProvider>
     </Provider>
   )