+++ packages/api-explorer/src/components/DocMethodSummary/DocMethodSummary.tsx
@@ -27,7 +27,7 @@ import React, { FC } from 'react'
 import { Card, CardContent, Flex, Space, Text } from '@looker/components'
 import { IMethod } from '@looker/sdk-codegen'
 import { MethodBadge } from '@looker/run-it'
-import { ApixHeading } from '../common'
+import { RunItHeading } from '../common'
 import { DocActivityType, DocRateLimited } from '../../components'
 import { DocSummaryStatus } from './DocSummaryStatus'
 
@@ -47,9 +47,9 @@ export const DocMethodSummary: FC<DocMethodSummaryProps> = ({ method }) => (
           {method.httpMethod.toUpperCase()}
         </MethodBadge>
         <Flex alignItems="start" flexDirection="column" flex="1" mx="small">
-          <ApixHeading as="h3" mb="0" pt="0">
+          <RunItHeading as="h3" mb="0" pt="0">
             {method.summary}
-          </ApixHeading>
+          </RunItHeading>
           <Space>
             <DocRateLimited method={method} />
             <DocActivityType method={method} />