+++ packages/api-explorer/src/components/common/index.ts
@@ -23,4 +23,5 @@
  SOFTWARE.
 
  */
-export { ApixHeading, ApixSection } from './common'
+export { RunItHeading, ApixSection } from './common'
+export { Loader } from './Loader'