+++ packages/run-it/src/components/DocSdkCalls/CodeCopy.tsx
@@ -23,28 +23,36 @@
  SOFTWARE.
 
  */
-import React from 'react'
-import { renderWithTheme } from '@looker/components-test-utils'
-import { Text } from '@looker/components'
-import { screen, waitFor } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
+import React, { FC } from 'react'
+import { CodeDisplay, CodeDisplayProps } from '@looker/code-editor'
+import { Space, CopyToClipboard } from '@looker/components'
 
-import { CollapserCard } from './CollapserCard'
-
-describe('CollapserCard', () => {
-  test('it renders heading, children and collapses', async () => {
-    const heading = 'Foo'
-    const childText = 'Bar'
-    renderWithTheme(
-      <CollapserCard heading={heading}>
-        <Text>{childText}</Text>
-      </CollapserCard>
-    )
-    const header = screen.getByText(heading)
-    expect(screen.getByText(childText)).toBeInTheDocument()
-    await userEvent.click(header)
-    await waitFor(() => {
-      expect(screen.queryByText(childText)).not.toBeInTheDocument()
-    })
-  })
-})
+interface CodeCopyProps extends CodeDisplayProps {
+  caption?: string
+}
+/**
+ * Shows code with clipboard copying support
+ */
+export const CodeCopy: FC<CodeCopyProps> = ({
+  language = 'json',
+  code,
+  caption = 'Copy',
+  pattern = '',
+  transparent = false,
+  inline = false,
+  lineNumbers = true,
+}) => {
+  return (
+    <Space between align="start" pt="xsmall">
+      <CodeDisplay
+        code={code}
+        language={language}
+        pattern={pattern}
+        transparent={transparent}
+        inline={inline}
+        lineNumbers={lineNumbers}
+      />
+      <CopyToClipboard content={code}>{caption}</CopyToClipboard>
+    </Space>
+  )
+}