+++ packages/run-it/src/components/RequestForm/RequestForm.tsx
@@ -26,9 +26,11 @@
 
 import React, { BaseSyntheticEvent, FC, Dispatch } from 'react'
 import { Button, Form, Space, ButtonTransparent } from '@looker/components'
-
+import type { IAPIMethods } from '@looker/sdk-rtl'
 import { RunItHttpMethod, RunItInput, RunItValues } from '../../RunIt'
 import { ConfigDialog, RunItConfigurator } from '../ConfigForm'
+import { LoginForm } from '../LoginForm'
+import { RunItSetter } from '../..'
 import {
   createSimpleItem,
   createComplexItem,
@@ -36,10 +38,11 @@ import {
   updateNullableProp,
 } from './formUtils'
 
-/**
- * Properties required by RequestForm
- */
+/** Properties required by RequestForm */
 interface RequestFormProps {
+  /** Established SDK instance */
+  sdk: IAPIMethods
+  /** Request inputs to the endpoint */
   inputs: RunItInput[]
   /** A callback for submitting the form */
   handleSubmit: (e: BaseSyntheticEvent) => void
@@ -48,7 +51,13 @@ interface RequestFormProps {
   /** A collection type react state to store path, query and body parameters as entered by the user  */
   requestContent: RunItValues
   /** A set state callback fn for populating requestContent on interaction with the request form */
-  setRequestContent: Dispatch<{ [key: string]: any }>
+  setRequestContent: Dispatch<RunItValues>
+  /** Is authentication required? */
+  needsAuth: boolean
+  /** Does RunIt have the configuration values it needs? */
+  hasConfig: boolean
+  /** Hook to refresh specifications */
+  setVersionsUrl: RunItSetter
   /** A set state callback which if present allows for editing, setting or clearing OAuth configuration parameters */
   setHasConfig?: Dispatch<boolean>
   /** Configuration plug-in for stand-alone or extension */
@@ -62,11 +71,15 @@ interface RequestFormProps {
  * inputs
  */
 export const RequestForm: FC<RequestFormProps> = ({
+  sdk,
   inputs,
   httpMethod,
   handleSubmit,
   requestContent,
   setRequestContent,
+  needsAuth,
+  hasConfig,
+  setVersionsUrl,
   setHasConfig,
   configurator,
   isExtension = false,
@@ -123,12 +136,27 @@ export const RequestForm: FC<RequestFormProps> = ({
         <ButtonTransparent type="button" onClick={handleClear}>
           Clear
         </ButtonTransparent>
-        <Button type="submit">Run</Button>
-        {!isExtension && setHasConfig && (
-          <ConfigDialog
-            setHasConfig={setHasConfig}
-            configurator={configurator}
-          />
+        {hasConfig ? (
+          needsAuth ? (
+            <LoginForm
+              sdk={sdk}
+              setVersionsUrl={setVersionsUrl}
+              setHasConfig={setHasConfig}
+              configurator={configurator}
+              requestContent={requestContent}
+            />
+          ) : (
+            <Button type="submit">Run</Button>
+          )
+        ) : (
+          !isExtension &&
+          setHasConfig && (
+            <ConfigDialog
+              setHasConfig={setHasConfig}
+              configurator={configurator}
+              setVersionsUrl={setVersionsUrl}
+            />
+          )
         )}
       </Space>
     </Form>